Add PostgreSQL backend support alongside SQLite (#48)
* Add PostgreSQL backend support alongside SQLite

The job_history database layer previously assumed SQLite exclusively.
This commit makes the backend switchable via environment variable, with
SQLite remaining the default and all existing behaviour unchanged.

## New files

job_history/config.py
  JobHistoryConfig class (modelled on SAMConfig from project_samuel).
  Loads a .env file at import time via python-dotenv, then exposes
  DB_BACKEND, SQLITE_DATA_DIR, PG_HOST/PORT/USER/PASSWORD/REQUIRE_SSL,
  and pg_db_name(machine) as class-level attributes.  Per-machine
  database names default to {machine}_jobs (derecho_jobs, casper_jobs).
  validate_postgres() fails fast if credentials are missing.

.env.example
  Documents all JH_* and legacy JOB_HISTORY_DATA_DIR / QHIST_*_DB
  variables.  Copy to .env and fill in; .env is gitignored.

.github/workflows/hpc-usage-queries-ci-postgres.yaml
  Second CI workflow that spins up a postgres:18 service container and
  runs the full test suite (pytest + integration test) with
  JH_DB_BACKEND=postgres.  Mirrors the existing SQLite workflow exactly
  except for the service block, job-level env vars, and the
  pip install .[dev,postgres] extras group.

## Modified files

job_history/database.py
  - Imports JobHistoryConfig; JOB_HISTORY_DATA_DIR kept as a
    backward-compat alias pointing at config.SQLITE_DATA_DIR.
  - get_db_url(machine): new backend-agnostic helper that returns a
    SQLite path string or a redacted PostgreSQL URL; used for display.
  - get_db_path(machine): now raises RuntimeError when backend != sqlite
    so callers learn they need get_db_url() instead.
  - get_engine(): branches on DB_BACKEND.  SQLite path is unchanged
    (PRAGMA listener, WAL, etc.).  Postgres path builds a
    postgresql+psycopg2 URL, honours PG_REQUIRE_SSL, attaches no PRAGMA.
  - init_db(): for postgres, calls _ensure_pg_database() first, which
    connects to the 'postgres' maintenance DB with AUTOCOMMIT isolation
    and issues CREATE DATABASE if the target DB is absent.
  - _set_sqlite_pragma / set_sqlite_pragma alias preserved unchanged.

job_history/sync_cli/sync.py
  Replaces get_db_path(machine) (verbose display only) with get_db_url()
  so the message is meaningful for both backends.

job_history/__init__.py
  Exports get_db_url and JobHistoryConfig alongside existing symbols.

job_history/sync.py
  Removes the long-unused 'from sqlalchemy.dialects.sqlite import insert
  as sqlite_insert' import.

job_history/tests/integtation_test.sh
  Backend-aware: reads JH_DB_BACKEND (default sqlite).  SQLite branch is
  unchanged (temp dir, ls *.db, rm -rf).  Postgres branch skips the temp
  dir, prints connection URLs after init_db(), and prints job/row counts
  from each database at the end instead of file sizes.

pyproject.toml
  Adds python-dotenv to core dependencies; adds [postgres] optional
  extras group containing psycopg2-binary.

conda-env.yaml
  Adds python-dotenv and psycopg2-binary under pip dependencies.

compose.yaml
  Corrects the service name to 'postgres' (was mislabelled 'mysql'),
  removes leftover MySQL environment variables, fixes the volume mount
  path, and adds a pg_isready healthcheck.


* Fix cputime overflow on PostgreSQL: Integer → BigInteger

SQLite stores integers with arbitrary precision, so this was silently
fine there.  PostgreSQL INTEGER is strictly 32-bit (max 2,147,483,647).

For large Derecho jobs, resources_used.cput in PBS accounting logs
is the *total* CPU-seconds summed across all ranks.  A 20,000-CPU job
running for 30 h produces 20,000 × 30 × 3600 = 2.16 B CPU-seconds,
which overflows a 32-bit column.  Full-machine Derecho jobs at
~318 k CPUs overflow after just ~7 h of walltime.

Fix: change cputime from Integer to BigInteger in the Job model.
reqmem / memory / vmemory were already BigInteger (correct).
elapsed and walltime stay Integer (seconds of walltime, max ~86 400
for a 24 h job — well within 32-bit range).


* Gitignore .env; fix compose volume path

.gitignore: add .env so credentials are never accidentally committed.
compose.yaml: correct postgres volume mount path (drop /data suffix
which caused the container to store data one level deeper than expected).


---------

+++ job_history/models.py
@@ -252,7 +252,7 @@ class Job(LookupMixin, Base):
     # Time metrics (in seconds)
     elapsed = Column(Integer)
     walltime = Column(Integer)
-    cputime = Column(Integer)
+    cputime = Column(BigInteger)  # total CPU-seconds; can exceed 2^31 for large parallel jobs
 
     # Resource allocation
     numcpus = Column(Integer)
